nfsd: stricter decoding of write-like NFSv2/v3 ops
Issue: SCP6-978

1/3 [
Author: Catalin Enache
Email: catalin.enache@windriver.com
Subject: svcrdma: Do not add XDR padding to xdr_buf page vector
Date: Wed, 17 May 2017 08:03:17 +0300

commit 6625d0913771df5f12b9531c8cb8414e55f1c21d upstream

An xdr_buf has a head, a vector of pages, and a tail. Each
RPC request is presented to the NFS server contained in an
xdr_buf.

The RDMA transport would like to supply the NFS server with only
the NFS WRITE payload bytes in the page vector. In some common
cases, that would allow the NFS server to swap those pages right
into the target file's page cache.

Have the transport's RDMA Read logic put XDR pad bytes in the tail
iovec, and not in the pages that hold the data payload.

The NFSv3 WRITE XDR decoder is finicky about the lengths involved,
so make sure it is looking in the correct places when computing
the total length of the incoming NFS WRITE request.

Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]

2/3 [
Author: Catalin Enache
Email: catalin.enache@windriver.com
Subject: nfsd4: minor NFSv2/v3 write decoding cleanup
Date: Wed, 17 May 2017 08:24:37 +0300

commit db44bac41bbfc0c0d9dd943092d8bded3c9db19b upstream

Use a couple shortcuts that will simplify a following bugfix.

Cc: stable@vger.kernel.org
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]

3/3 [
Author: "J. Bruce Fields"
Email: bfields@redhat.com
Subject: nfsd: stricter decoding of write-like NFSv2/v3 ops
Date: Fri, 21 Apr 2017 15:26:30 -0400

commit 13bf9fbff0e5e099e2b6f003a0ab8ae145436309 upstream

The NFSv2/v3 code does not systematically check whether we decode past
the end of the buffer.  This generally appears to be harmless, but there
are a few places where we do arithmetic on the pointers involved and
don't account for the possibility that a length could be negative.  Add
checks to catch these.

Reported-by: Tuomas Haanpää <thaan@synopsys.com>
Reported-by: Ari Kauppi <ari@synopsys.com>
Reviewed-by: NeilBrown <neilb@suse.com>
Cc: stable@vger.kernel.org
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>