arm64: use fixmap region for permanent FDT mapping
Issue: LIN8-6639

1/2 [
Author: Ard Biesheuvel
Email: ard.biesheuvel@linaro.org
Subject: of/fdt: split off FDT self reservation from memreserve processing
Date: Mon, 1 Jun 2015 13:40:31 +0200

commit 24bbd929e6b9e62afd263c42b4318d3b603c956c upstream.

This splits off the reservation of the memory occupied by the FDT
binary itself from the processing of the memory reservations it
contains. This is necessary because the physical address of the FDT,
which is needed to perform the reservation, may not be known to the
FDT driver core, i.e., it may be mapped outside the linear direct
mapping, in which case __pa() returns a bogus value.

Cc: Russell King <linux@arm.linux.org.uk>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Cc: Paul Mackerras <paulus@samba.org>
Acked-by: Rob Herring <robh@kernel.org>
Acked-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Li Zhou <li.zhou@windriver.com>
]

2/2 [
Author: Ard Biesheuvel
Email: ard.biesheuvel@linaro.org
Subject: arm64: use fixmap region for permanent FDT mapping
Date: Mon, 1 Jun 2015 13:40:32 +0200

commit 61bd93ce801bb6df36eda257a9d2d16c02863cdd upstream.

Currently, the FDT blob needs to be in the same 512 MB region as
the kernel, so that it can be mapped into the kernel virtual memory
space very early on using a minimal set of statically allocated
translation tables.

Now that we have early fixmap support, we can relax this restriction,
by moving the permanent FDT mapping to the fixmap region instead.
This way, the FDT blob may be anywhere in memory.

This also moves the vetting of the FDT to mmu.c, since the early
init code in head.S does not handle mapping of the FDT anymore.
At the same time, fix up some comments in head.S that have gone stale.

Reviewed-by: Mark Rutland <mark.rutland@arm.com>
Tested-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Li Zhou <li.zhou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>