drm/vmwgfx: Make sure backup_handle is always valid
Issue: LIN9-4496

1/1 [
Author: Sinclair Yeh
Email: syeh@vmware.com
Subject: drm/vmwgfx: Make sure backup_handle is always valid
Date: Fri, 2 Jun 2017 07:50:57 +0200

commit 07678eca2cf9c9a18584e546c2b2a0d0c9a3150c upstream

When vmw_gb_surface_define_ioctl() is called with an existing buffer,
we end up returning an uninitialized variable in the backup_handle.

The fix is to first initialize backup_handle to 0 just to be sure, and
second, when a user-provided buffer is found, we will use the
req->buffer_handle as the backup_handle.

Cc: <stable@vger.kernel.org>
Reported-by: Murray McAllister <murray.mcallister@insomniasec.com>
Signed-off-by: Sinclair Yeh <syeh@vmware.com>
Reviewed-by: Deepak Rawat <drawat@vmware.com>
Signed-off-by: Haiqing Bai <Haiqing.Bai@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>