Make asm/word-at-a-time.h available on all architectures
Issue: LIN8-7433

1/1 [
Author: Chris Metcalf
Email: cmetcalf@ezchip.com
Subject: Make asm/word-at-a-time.h available on all architectures
Date: Wed, 29 Apr 2015 12:48:40 -0400

commit a6e2f029ae34f41adb6ae3812c32c5d326e1abd2 upstream

Added the x86 implementation of word-at-a-time to the
generic version, which previously only supported big-endian.

Omitted the x86-specific load_unaligned_zeropad(), which in
any case is also not present for the existing BE-only
implementation of a word-at-a-time, and is only used under
CONFIG_DCACHE_WORD_ACCESS.

Added as a "generic-y" to the Kbuilds of all architectures
that didn't previously have it.

Signed-off-by: Chris Metcalf <cmetcalf@ezchip.com>
[Cherry-picked from upstream for mips]
Signed-off-by: Zhixiong Chi <zhixiong.chi@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>