mnt: Fail collect_mounts when applied to unmounted mounts
Issue: LIN6-11199

1/1 [
Author: "Eric W. Biederman"
Email: ebiederm@xmission.com
Subject: mnt: Fail collect_mounts when applied to unmounted mounts
Date: Wed, 7 Jan 2015 14:28:26 -0600

The only users of collect_mounts are in audit_tree.c

commit cd4a40174b71acd021877341684d8bb1dc8ea4ae upstream

In audit_trim_trees and audit_add_tree_rule the path passed into
collect_mounts is generated from kern_path passed an audit_tree
pathname which is guaranteed to be an absolute path.   In those cases
collect_mounts is obviously intended to work on mounted paths and
if a race results in paths that are unmounted when collect_mounts
it is reasonable to fail early.

The paths passed into audit_tag_tree don't have the absolute path
check.  But are used to play with fsnotify and otherwise interact with
the audit_trees, so again operating only on mounted paths appears
reasonable.

Avoid having to worry about what happens when we try and audit
unmounted filesystems by restricting collect_mounts to mounts
that appear in the mount tree.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
[The commit fix CVE-2015-4177]
Signed-off-by: Zhenbo Gao <zhenbo.gao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>