kvm-rt: drop patch adding manual override of lapic irqsafe flag
Issue: OVP9-37

We originally added this when debugging interrupt latency regressions
on the FPGA test several releases ago -- as a convenience aid.

We are no longer doing the FPGA test, and further, if you inspect
mainline kernels of today, the kvm_create_lapic() doesn't even
have this flag anymore.  So we drop this patch.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>