kvm-rt: drop patch adding "tsc=perfect" option
Issue: OVP9-37

There have been a lot of changes in the code around TSC stability
detection, and dealing with cores/systems that have issues, in the
five years since this patch was written.

Given that, I think this might be a case where "user knows best"
may not apply, and we've also seen people blindly copy boot args
from one instance to another which backs up that statement.

We drop this patch and limit ourselves to the tsc= bootargs set
that the mainline kernel supports, assuming they are limited
in scope to prevent people hurting themselves too badly.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>