kvm-rt: drop async page fault lock conversion patch
Issue: OVP9-37

I created this change four years ago on a 3.4.x kernel since
the trace showed the apf_wake_all was being called from atomic
context and using locks unsuitable for that context.

However, I think this pre-dated the patch to convert the APF
to use a simple wait queue, which has been in -rt for quite
some time (years) and is even now mainline (4.6+)

So, rather than death march this change forward again, I'd like
to see that it is still an issue.

If it was a corner case issue only seen sporadically, then I'd
be less inclined to take this approach.  However, we can see
from the trace that it happened very early, pretty much right
after a CPU that was idle since boot finally took a page fault.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>