flow_dissector: Jump to exit code in __skb_flow_dissect
Issue: SCP6-1030

1/1 [
Author: Ovidiu Panait
Email: ovidiu.panait@windriver.com
Subject: flow_dissector: Jump to exit code in __skb_flow_dissect
Date: Sat, 9 Sep 2017 16:41:35 +0000

commit a6e544b0a88b53114bfa5a57e21b7be7a8dfc9d0 upstream

Instead of returning immediately (on a parsing failure for instance) we
jump to cleanup code. This always sets protocol values in key_control
(even on a failure there is still valid information in the key_tags that
was set before the problem was hit).

Signed-off-by: Tom Herbert <tom@herbertland.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>