fix kernel crash caused by CVE-2017-7533
Issue: SCP7-664

1/1 [
Author: Liu Haitao
Email: haitao.liu@windriver.com
Subject: fix kernel crash caused by CVE-2017-7533
Date: Thu, 14 Sep 2017 01:34:26 -0700

Issue: LIN7-8722 LIN7-8735

When booting up the target, it fails with:

```

	[<ffffffff8117bf64>] SyS_renameat+0xc4/0x390
	[<ffffffff811ac58c>] ? fsnotify+0x27c/0x350
	[<ffffffff8116ab0a>] ? do_sync_write+0x5a/0x90
	[<ffffffff8116b2aa>] ? vfs_write+0x17a/0x1b0
	[<ffffffff8117c24b>] SyS_rename+0x1b/0x20
	[<ffffffff81825ff6>] system_call_fastpath+0x1a/0x1f

```
There is a mistake when backporint the upstream commit to fix CVE-2017-7533.
The 'take_dentry_name_snapshot' do not consider the bound of array.
When tackling the file with a long filename this function would cause memory
overflow issue.

And in linux kernel, if a filename consist of only a few characters, it is
held in 'struct dentry -> d_iname'. But if it has many characters it is held in
"struct dentry -> d_name" instead of "d_iname".

Signed-off-by: Haitao Liu <Haitao.Liu@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>