fix kernel crash caused by CVE-2017-7533
Issue: SCP6-1037

1/1 [
Author: Liu Haitao
Email: haitao.liu@windriver.com
Subject: fix kernel crash caused by CVE-2017-7533
Date: Thu, 14 Sep 2017 17:59:10 +0800

Issue: LIN6-13736

When renaming a file with a long filename, it fails with:

```
	[<ffffffff8118da51>] ? vfsmount_lock_local_unlock+0x31/0x60
	[<ffffffff8116f3a8>] ? fput+0x1a8/0x250
	[<ffffffff8116b1e6>] ? filp_close+0x66/0x90
	[<ffffffff8100c818>] ? pax_randomize_kstack+0x38/0x70
	[<ffffffff818822ee>] ? sysret_check+0x17/0x53
	[<ffffffff8117f93b>] sys_rename+0x1b/0x20
	[<ffffffff818822cd>] system_call_fastpath+0x1c/0x21

```

There is a mistake when backporint the upstream commit to fix CVE-2014-7533.
The 'take_dentry_name_snapshot' do not consider the bound of array.
When tackling the file with a long filename this function would cause memory
overflow issue.

And in linux kernel, if a filename consist of only a few characters, it is
held in 'struct dentry -> d_iname'. But if it has many characters it is held in
"struct dentry -> d_name" instead of "d_iname".

Signed-off-by: Haitao Liu <Haitao.Liu@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>