Revert "AHCI: Added MSIX interrupt support to SATA PCI driver"
Issue: LIN8-8061

1/1 [
Author: Yanjiang Jin
Email: yanjiang.jin@windriver.com
Subject: Revert "AHCI: Added MSIX interrupt support to SATA PCI driver"
Date: Wed, 25 Oct 2017 15:16:38 +0800

This patch should be for Cavium Thunder only, which is an ARM64 board.
It shouldn't be for Cavium Octeon, which is a MIPS board.
The macro 'AHCI_HFLAG_MSIX' is only defined in Cavium Thunder environment,
it would lead to a build error.
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>