driver: spi-nor: remove warning info from spi-nor driver
Issue: LIN9-5827

1/1 [
Author: Limeng
Email: Meng.Li@windriver.com
Subject: driver: spi-nor: remove warning info from spi-nor driver
Date: Thu, 16 Nov 2017 23:10:13 +0800

Remove the warning info that is outputted by WARN_ONCE that
will cause a warning trace when offset is not 0.
In additional, data may be written into any page of any block, this
is an normal case. So, it is not need to show warning info.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>