driver : usb: dwc3: there is no need to report a warning trace
Issue: LIN9-5845

1/1 [
Author: Limeng
Email: Meng.Li@windriver.com
Subject: driver : usb: dwc3: there is no need to report a warning trace
Date: Fri, 25 Aug 2017 16:55:41 +0800

In function dwc3_frame_length_adjustment(), if the frame length
value that will be set into register is the same with the value in
register, there will a warning trace is reported. It is a little
making a fuss, so change it to a normal warning and report it with
function dev_warn().

Signed-off-by: Meng Li <Meng.Li@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>