crypto: ccam - remove clrsetbits_be32 and clrsetbits_le32
Issue: LIN8-8045

1/1 [
Author: Liu Haitao
Email: haitao.liu@windriver.com
Subject: crypto: ccam - remove clrsetbits_be32 and clrsetbits_le32
Date: Wed, 20 Dec 2017 00:07:09 -0800

There are some kernel warnings of duplicate definition  When build image for
arm.
The clrsetbits_be32 and clrsetbits_le32 have been defined in "include/asm/io.h".

So they should not be defined again in "ccam/regs.h".

The standard do not have this issue, for the following commit have indirectly
fixed it.

"""
    261ea058f016bc04fa064348ad9bf39d94379381
"""

As this commit is complex and inapplicable for the current kernel, we just adopt
a simple approach to fix it.

Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>