crypto: pcrypt - fix freeing pcrypt instances
Issue: LIN9-6302

1/1 [
Author: Eric Biggers
Email: ebiggers@google.com
Subject: crypto: pcrypt - fix freeing pcrypt instances
Date: Wed, 20 Dec 2017 14:28:25 -0800

commit d76c68109f37cb85b243a1cf0f40313afd2bae68 upstream

pcrypt is using the old way of freeing instances, where the ->free()
method specified in the 'struct crypto_template' is passed a pointer to
the 'struct crypto_instance'.  But the crypto_instance is being
kfree()'d directly, which is incorrect because the memory was actually
allocated as an aead_instance, which contains the crypto_instance at a
nonzero offset.  Thus, the wrong pointer was being kfree()'d.

Fix it by switching to the new way to free aead_instance's where the
->free() method is specified in the aead_instance itself.

Reported-by: syzbot <syzkaller@googlegroups.com>
Fixes: 0496f56065e0 ("crypto: pcrypt - Add support for new AEAD interface")
Cc: <stable@vger.kernel.org> # v4.2+
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>