mtd: spi-nor: remove WARN_ONCE() message in spi_nor_write()
Issue: LIN9-6880

1/1 [
Author: Cyrille Pitchen
Email: cyrille.pitchen@atmel.com
Subject: mtd: spi-nor: remove WARN_ONCE() message in spi_nor_write()
Date: Tue, 6 Dec 2016 18:14:24 +0100

commit bc0e151514d09cadb56e473a10c783e64e48ce0b upstream

This patch removes the WARN_ONCE() test in spi_nor_write().
This macro triggers the display of a warning message almost every time we
use a UBI file-system because a write operation is performed at offset 64,
which is in the middle of the SPI NOR memory page. This is a valid
operation for ubifs.

Hence this warning is pretty annoying and useless so we just remove it.

Signed-off-by: Cyrille Pitchen <cyrille.pitchen@atmel.com>
Suggested-by: Richard Weinberger <richard@nod.at>
Suggested-by: Andras Szemzo <szemzo.andras@gmail.com>
Acked-by: Boris Brezillon <boris.brezillon@free-electrons.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>