cdrom: information leak in cdrom_ioctl_media_changed()
Issue: SCP6-1192

1/1 [
Author: Dan Carpenter
Email: dan.carpenter@oracle.com
Subject: cdrom: information leak in cdrom_ioctl_media_changed()
Date: Wed, 18 Apr 2018 12:51:31 +0300

commit 9de4ee40547fd315d4a0ed1dd15a2fa3559ad707 upstream

This cast is wrong.  "cdi->capacity" is an int and "arg" is an unsigned
long.  The way the check is written now, if one of the high 32 bits is
set then we could read outside the info->slots[] array.

This bug is pretty old and it predates git.

Reviewed-by: Christoph Hellwig <hch@lst.de>
Cc: stable@vger.kernel.org
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Li Zhou <li.zhou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>