KVM: x86: fix singlestepping over syscall
Issue: LIN8-9334

1/1 [
Author: Paolo Bonzini
Email: pbonzini@redhat.com
Subject: KVM: x86: fix singlestepping over syscall
Date: Tue, 14 Aug 2018 15:18:39 +0800

commit c8401dda2f0a00cd25c0af6a95ed50e478d25de4 upstream

TF is handled a bit differently for syscall and sysret, compared
to the other instructions: TF is checked after the instruction completes,
so that the OS can disable #DB at a syscall by adding TF to FMASK.
When the sysret is executed the #DB is taken "as if" the syscall insn
just completed.

KVM emulates syscall so that it can trap 32-bit syscall on Intel processors.
Fix the behavior, otherwise you could get #DB on a user stack which is not
nice.  This does not affect Linux guests, as they use an IST or task gate
for #DB.

This fixes CVE-2017-7518.

Cc: stable@vger.kernel.org
Reported-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Radim Krčmář <rkrcmar@redhat.com>
[Change or drop some codes to apply it into our tree.]
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>