hfsplus: fix NULL dereference in hfsplus_lookup()
Issue: LIN7-10022

1/1 [
Author: Liu Haitao
Email: haitao.liu@windriver.com
Subject: hfsplus: fix NULL dereference in hfsplus_lookup()
Date: Mon, 13 Aug 2018 20:01:53 -0700

commit bfa2d20c143e1a5b7203195ad81d462d45a6961d upstream

An HFS+ filesystem can be mounted read-only without having a metadata
directory, which is needed to support hardlinks. But if the catalog data
is corrupted, a directory lookup may still find dentries claiming to be
hardlinks.

hfsplus_lookup() does check that ->hidden_dir is not NULL in such a situation,
but mistakenly does so after dereferencing it for the first time. Reorder
this check to prevent a crash.

This happens when looking up corrupted catalog data (dentry) on a filesystem with
no metadata directory (this could only ever happen on a read-only mount). Wen Xu
sent the replication steps in detail to the fsdevel list:
https://bugzilla.kernel.org/show_bug.cgi?id=200297

Link: http://lkml.kernel.org/r/20180712215344.q44dyrhymm4ajkao@eaf
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>
Reported-by: Wen Xu <wen.xu@gatech.edu>
Cc: Viacheslav Dubeyko <slava@dubeyko.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>
[hl:update the patch to adjust for 3.10 and 3.14 kernel context]
Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>