ext4: avoid running out of journal credits when appending to an inline file
Issue: LIN9-7204

1/2 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: jbd2: don't mark block as modified if the handle is out of credits
Date: Sat, 16 Jun 2018 20:21:45 -0400

commit e09463f220ca9a1a1ecfda84fcda658f99a1f12a upstream

Do not set the b_modified flag in block's journal head should not
until after we're sure that jbd2_journal_dirty_metadat() will not
abort with an error due to there not being enough space reserved in
the jbd2 handle.

Otherwise, future attempts to modify the buffer may lead a large
number of spurious errors and warnings.

This addresses CVE-2018-10883.

https://bugzilla.kernel.org/show_bug.cgi?id=200071

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@kernel.org
Signed-off-by: Rui Wang <rui.wang@windriver.com>
]

2/2 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: avoid running out of journal credits when appending to an inline file
Date: Sat, 16 Jun 2018 23:41:59 -0400

commit 8bc1379b82b8e809eef77a9fedbb75c6c297be19 upstream

Use a separate journal transaction if it turns out that we need to
convert an inline file to use an data block.  Otherwise we could end
up failing due to not having journal credits.

This addresses CVE-2018-10883.

https://bugzilla.kernel.org/show_bug.cgi?id=200071

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Cc: stable@kernel.org
Signed-off-by: Rui Wang <rui.wang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>