ext4: always verify the magic number in xattr blocks
Issue: SCP6-1262

1/2 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: add corruption check in ext4_xattr_set_entry()
Date: Wed, 13 Jun 2018 00:23:11 -0400

commit 5369a762c882c0b6e9599e4ebbb3a9ba9eee7e2d upstream

In theory this should have been caught earlier when the xattr list was
verified, but in case it got missed, it's simple enough to add check
to make sure we don't overrun the xattr buffer.

This addresses CVE-2018-10879.

https://bugzilla.kernel.org/show_bug.cgi?id=200001

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Cc: stable@kernel.org
Signed-off-by: Rui Wang <rui.wang@windriver.com>
]

2/2 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: ext4: always verify the magic number in xattr blocks
Date: Wed, 13 Jun 2018 00:51:28 -0400

commit 513f86d73855ce556ea9522b6bfd79f87356dc3a upstream

If there an inode points to a block which is also some other type of
metadata block (such as a block allocation bitmap), the
buffer_verified flag can be set when it was validated as that other
metadata block type; however, it would make a really terrible external
attribute block.  The reason why we use the verified flag is to avoid
constantly reverifying the block.  However, it doesn't take much
overhead to make sure the magic number of the xattr block is correct,
and this will avoid potential crashes.

This addresses CVE-2018-10879.

https://bugzilla.kernel.org/show_bug.cgi?id=200001

Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Cc: stable@kernel.org
Signed-off-by: Rui Wang <rui.wang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>