cdrom: fix improper type cast, which can leat to information leak
Issue: SCP7-939

1/1 [
Author: Li Zhou
Email: li.zhou@windriver.com
Subject: cdrom: fix improper type cast, which can leat to information leak.
Date: Thu, 1 Nov 2018 13:14:17 +0800

commit e4f3aa2e1e67bb48dfbaaf1cad59013d5a5bc276 upstream

There is another cast from unsigned long to int which causes
a bounds check to fail with specially crafted input. The value is
then used as an index in the slot array in cdrom_slot_status().

This issue is similar to CVE-2018-16658 and CVE-2018-10940.

Signed-off-by: Young_X <YangX92@hotmail.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Li Zhou <li.zhou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>