stmmac: fix resource management when resume
Issue: LIN8-10408

1/1 [
Author: Ovidiu Panait
Email: ovidiu.panait@windriver.com
Subject: stmmac: fix resource management when resume
Date: Tue, 5 Feb 2019 09:04:16 +0200

commit ae79a639bb3dfd168dc8c1e5d6dfc471bdf6f284 upstream

There is a memleak when suspend/resume this driver version.
Currently the stmmac, during resume step, reallocates all the resources
but they are not released when suspend.
The patch is not to release these resources but the logic has been changed.
In fact, it is not necessary to free and reallocate all from scratch
because the memory data will be always preserved.
As final solution, the patch just reinit the descriptors and the rx/tx
pointers only when resume. Tested done on STi boxes.

Reported-by: ZhengShunQian <zhengsq@rock-chips.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Cc: David S. Miller <davem@davemloft.net>
Signed-off-by: David S. Miller <davem@davemloft.net>
[OP: adjusted context for kernel version 4.1, Issue: LIN8-10408]
Signed-off-by: Ovidiu Panait <ovidiu.panait@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>