arm64: make sys_call_table const
Issue: SCP7-494

1/1 [
Author: Mark Rutland
Email: mark.rutland@arm.com
Subject: arm64: make sys_call_table const
Date: Thu, 8 Jan 2015 11:42:59 +0000

commit c623b33b4e9599c6ac5076f7db7369eb9869aa04 upstream

As with x86, mark the sys_call_table const such that it will be placed
in the .rodata section. This will cause attempts to modify the table
(accidental or deliberate) to fail when strict page permissions are in
place. In the absence of strict page permissions, there should be no
functional change.

Signed-off-by: Mark Rutland <mark.rutland@arm.com>
Acked-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>