Failing to send a CLOSE if file is opened WRONLY and server reboots on a 4.x mount
Issue: LIN8-10773

1/1 [
Author: Olga Kornievskaia
Email: aglo@umich.edu
Subject: Failing to send a CLOSE if file is opened WRONLY and server reboots on a 4.x mount
Date: Mon, 14 Sep 2015 19:54:36 -0400

commit a41cbe86df3afbc82311a1640e20858c0cd7e065 upstream

A test case is as the description says:
open(foobar, O_WRONLY);
sleep()  --> reboot the server
close(foobar)

The bug is because in nfs4state.c in nfs4_reclaim_open_state() a few
line before going to restart, there is
clear_bit(NFS4CLNT_RECLAIM_NOGRACE, &state->flags).

NFS4CLNT_RECLAIM_NOGRACE is a flag for the client states not open
owner states. Value of NFS4CLNT_RECLAIM_NOGRACE is 4 which is the
value of NFS_O_WRONLY_STATE in nfs4_state->flags. So clearing it wipes
out state and when we go to close it, “call_close” doesn’t get set as
state flag is not set and CLOSE doesn’t go on the wire.

Signed-off-by: Olga Kornievskaia <aglo@umich.edu>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>